requirements.txt
keras
tensorflow
numpy
flask
pillow
